Update class InfixExpressionToPolishExpressionConverter
